style: 구름 컴포넌트 분리
